+++ src/php/Gdbots/Schemas/Common/Mixin/Taggable/Taggable.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Gdbots\Schemas\Common\Mixin\Taggable;
+
+use Gdbots\Pbj\Message;
+
+interface Taggable extends Message
+{
+}